% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicvisualizer.R
\name{musicvisualizer}
\alias{musicvisualizer}
\title{Retrieve metadata relevant to the albums of an artist}
\usage{
musicvisualizer(artist_n, album_n)
}
\arguments{
\item{artist_n}{A string character vector identifying an existing singer that possesses the album}

\item{album_n}{A string character vector identifying an existing album}
}
\value{
A data frame of class \code{musicvisualizer} with the following fields
\itemize{
\item track_name
\item danceability
\item energy
\item valence
\item instrumentalness
\item acousticness
}

Note
}
\description{
Given an artist's name, this function retrieves a dataset from Spotify that
includes relevant albums' information such as danceability, energy, valence, etc.
}
\examples{
first_musicvisualizer <- musicvisualizer("Olivia Rodrigo", "GUTS")

}
