% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicvisualizer.R
\name{plot.musicvisualizer}
\alias{plot.musicvisualizer}
\title{Visualize the relevant features - danceability, energy, valence, instrumentalness, and acousticness}
\usage{
\method{plot}{musicvisualizer}(x, ...)
}
\arguments{
\item{x}{An \link{'musicvisualizer'} object}

\item{...}{Currently ignored}
}
\description{
Description HERE!!! My brain is not FUNCTIONING at the moment!!! We will come back in one day!!!
}
