% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicvisualizer.R
\name{musicvisualizer}
\alias{musicvisualizer}
\title{Retrieve metadata relevant to the albums of an artist}
\usage{
musicvisualizer(album_name, artist_name)
}
\arguments{
\item{album_name}{A string character vector identifying an existing album}

\item{artist_name}{A string character vector identifying an existing singer that possesses the album}
}
\value{
A data frame of class \code{musicvisualizer} with the following fields
\itemize{
\item artist_name
\item album_images
\item album_release_date
\item danceability
\item energy
\item valence
\item instrumentalness
\item acousticness
\item track_name
\item album_name
}

Note
}
\description{
Given an artist's name, this function retrieves a dataset from Spotify that
includes relevant albums' information such as danceability, energy, valence, etc.
}
